Add other category to maker
